refactor: move some types to their own mods
