+++ crates/registry-mock/src/lib.rs
@@ -2,7 +2,11 @@ mod dirs;
 mod kill_verdaccio;
 mod mock_instance;
 mod node_registry_mock;
+mod registry_anchor;
+mod registry_info;
 
 pub use dirs::*;
 pub use mock_instance::*;
 pub use node_registry_mock::*;
+pub use registry_anchor::*;
+pub use registry_info::*;
